# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: List every platform user for admin management
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
        "401":
          description: Not signed in
        "403":
          description: Not an admin
  /admin/teachers:
    post:
      operationId: toggleAdminTeacher
      tags: [admin]
      summary: Promote or demote a user as teacher by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminToggleTeacherBody"
      responses:
        "200":
          description: Updated user role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not signed in
        "403":
          description: Not an admin
  /admin/classes:
    get:
      operationId: getAdminClasses
      tags: [admin]
      summary: List all classes with their teacher and students
      responses:
        "200":
          description: Classes list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminClassList"
        "401":
          description: Not signed in
        "403":
          description: Not an admin
    post:
      operationId: createAdminClass
      tags: [admin]
      summary: Create a new class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassBody"
      responses:
        "200":
          description: Created class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminClass"
        "401":
          description: Not signed in
        "403":
          description: Not an admin
  /admin/classes/{id}:
    patch:
      operationId: updateAdminClass
      tags: [admin]
      summary: Update a class name or teacher assignment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClassBody"
      responses:
        "200":
          description: Updated class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminClass"
        "401":
          description: Not signed in
        "403":
          description: Not an admin
    delete:
      operationId: deleteAdminClass
      tags: [admin]
      summary: Delete a class and detach its students
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Class deleted
        "401":
          description: Not signed in
        "403":
          description: Not an admin
        "404":
          description: Class not found
  /admin/students/{id}/class:
    patch:
      operationId: moveAdminStudentClass
      tags: [admin]
      summary: Move or remove a student from a class
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveStudentBody"
      responses:
        "200":
          description: Updated student class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStudentClass"
        "401":
          description: Not signed in
        "403":
          description: Not an admin
...
  /activity-logs:
    get:
      operationId: getActivityLogs
      tags: [activity]
      summary: Get activity logs for the signed-in student
      responses:
        "200":
          description: Activity logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityLogList"
        "401":
          description: Not signed in
  /admin/activity-logs/{studentId}:
    get:
      operationId: getAdminActivityLogs
      tags: [admin, activity]
      summary: Get activity logs for any student (admin only)
      parameters:
        - in: path
          name: studentId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activity logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityLogList"
        "401":
          description: Not signed in
        "403":
          description: Not an admin
...
components:
...
  schemas:
...
    AvatarConfig:
      type: object
      properties:
        bgColor:
          type: string
        accessories:
          type: array
          items:
            type: string
        gender:
          type: string
          enum: [male, female]
        pet:
          type: string
        nickname:
          type: string
          maxLength: 30
          default: ""
        frame:
          type: string
          default: "none"
        badges:
          type: array
          items:
            type: string
          default: []
      required:
        - bgColor
        - accessories
        - gender
        - pet
        - nickname
        - frame
        - badges
...
    AdminUser:
      type: object
      properties:
        studentId:
          oneOf:
            - type: integer
            - type: "null"
        replitUserId:
          oneOf:
            - type: string
            - type: "null"
        email:
          type: string
        name:
          type: string
        imageUrl:
          oneOf:
            - type: string
            - type: "null"
        role:
          type: string
          enum: [student, teacher]
        nameConfirmed:
          type: boolean
        classId:
          oneOf:
            - type: integer
            - type: "null"
        className:
          oneOf:
            - type: string
            - type: "null"
      required:
        - studentId
        - replitUserId
        - email
        - name
        - role
        - nameConfirmed
        - classId
        - className
    AdminUserList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
      required:
        - users
    AdminToggleTeacherBody:
      type: object
      properties:
        email:
          type: string
          minLength: 1
          maxLength: 320
        isTeacher:
          type: boolean
      required:
        - email
        - isTeacher
    AdminClass:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        teacherId:
          oneOf:
            - type: integer
            - type: "null"
        teacherName:
          oneOf:
            - type: string
            - type: "null"
        teacherEmail:
          oneOf:
            - type: string
            - type: "null"
        isChatEnabled:
          type: boolean
        students:
          type: array
          items:
            $ref: "#/components/schemas/TeacherStudent"
      required:
        - id
        - name
        - teacherId
        - teacherName
        - teacherEmail
        - isChatEnabled
        - students
    AdminClassList:
      type: object
      properties:
        classes:
          type: array
          items:
            $ref: "#/components/schemas/AdminClass"
      required:
        - classes
    CreateClassBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        teacherId:
          oneOf:
            - type: integer
            - type: "null"
      required:
        - name
    UpdateClassBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        teacherId:
          oneOf:
            - type: integer
            - type: "null"
        isChatEnabled:
          type: boolean
    MoveStudentBody:
      type: object
      properties:
        classId:
          oneOf:
            - type: integer
            - type: "null"
    AdminStudentClass:
      type: object
      properties:
        id:
          type: integer
        classId:
          oneOf:
            - type: integer
            - type: "null"
      required:
        - id
        - classId
...
    TeacherStudent:
      type: object
      properties:
        id:
          type: integer
        replitUserId:
          oneOf:
            - type: string
            - type: "null"
        name:
          type: string
        email:
          oneOf:
            - type: string
            - type: "null"
        points:
          type: integer
        avatarConfig:
          $ref: "#/components/schemas/AvatarConfig"
      required:
        - id
        - replitUserId
        - name
        - points
        - avatarConfig
...
    ActivityLogList:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/ActivityLog"
      required:
        - logs
    ActivityLog:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        type:
          type: string
        title:
          type: string
        description:
          type: string
        metadata:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - studentId
        - type
        - title
        - description
        - createdAt
...