# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUserRole } from './adminUserRole';

export interface AdminUser {
  studentId: number | null;
  replitUserId: string | null;
  email: string;
  name: string;
  imageUrl?: string | null;
  role: AdminUserRole;
  nameConfirmed: boolean;
  classId: number | null;
  className: string | null;
}