# lib/api-zod/src/generated/types/teacherStudent.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AvatarConfig } from './avatarConfig';

export interface TeacherStudent {
  id: number;
  replitUserId: string | null;
  name: string;
  email?: string | null;
  points: number;
  avatarConfig: AvatarConfig;
}